git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@2869 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/build/OnlineContent.java
@@ -141,7 +141,9 @@ public class OnlineContent {
 	private void moveProguardJar() throws IOException {
 		String source = deployDir + "/jsiddevice-" + projectVersion + "-proguard.jar";
 		String target = deployDir + "/jsiddevice-" + projectVersion + ".jar";
-		Files.move(Paths.get(source), Paths.get(target), REPLACE_EXISTING);
+		if (new File(source).exists()) {
+			Files.move(Paths.get(source), Paths.get(target), REPLACE_EXISTING);
+		}
 	}
 
 	private void upx() throws IOException, InterruptedException {
